Add files via upload
This project is an analysis of sales data for WA Company from 2004-2007, with a focus on identifying the most profitable countries and product lines. Using Tableau, the project includes visualizations of revenue by country, product line, and product type, as well as insights into average unit price, gross profit margin, and more. The goal of the project is to provide WA Company with data-driven insights to help streamline operations and increase profitability.